requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0 